# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AGRO TRADE marketplace API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: marketplace
    description: Consumer marketplace operations
  - name: farmers
    description: Farmer portal operations
  - name: orders
    description: Order and delivery operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /listings:
    get:
      operationId: listListings
      tags: [marketplace]
      summary: Browse crop listings
      parameters:
        - $ref: "#/components/parameters/CropQuery"
        - $ref: "#/components/parameters/LocationQuery"
        - $ref: "#/components/parameters/MinPriceQuery"
        - $ref: "#/components/parameters/MaxPriceQuery"
      responses:
        "200":
          description: Available crop listings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Listing"
    post:
      operationId: createListing
      tags: [farmers]
      summary: Create a crop listing
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ListingInput"
      responses:
        "201":
          description: Created listing
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Listing"
  /listings/{listingId}:
    get:
      operationId: getListing
      tags: [marketplace]
      summary: Get one listing
      parameters:
        - $ref: "#/components/parameters/ListingId"
      responses:
        "200":
          description: Listing detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Listing"
        "404":
          description: Listing not found
    patch:
      operationId: updateListing
      tags: [farmers]
      summary: Update a crop listing
      parameters:
        - $ref: "#/components/parameters/ListingId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ListingUpdate"
      responses:
        "200":
          description: Updated listing
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Listing"
  /listings/{listingId}/sold-out:
    post:
      operationId: markListingSoldOut
      tags: [farmers]
      summary: Mark a listing sold out
      parameters:
        - $ref: "#/components/parameters/ListingId"
      responses:
        "200":
          description: Sold out listing
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Listing"
  /farmers:
    get:
      operationId: listFarmers
      tags: [marketplace]
      summary: List farmers
      responses:
        "200":
          description: Farmer profiles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Farmer"
  /farmers/{farmerId}:
    get:
      operationId: getFarmer
      tags: [marketplace]
      summary: Get a farmer profile
      parameters:
        - $ref: "#/components/parameters/FarmerId"
      responses:
        "200":
          description: Farmer profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Farmer"
  /farmers/{farmerId}/listings:
    get:
      operationId: listFarmerListings
      tags: [marketplace]
      summary: List a farmer's crops
      parameters:
        - $ref: "#/components/parameters/FarmerId"
      responses:
        "200":
          description: Farmer listings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Listing"
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List orders
      parameters:
        - name: role
          in: query
          schema:
            type: string
            enum: [consumer, farmer]
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Place a household or bulk order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Created order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{orderId}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get order tracking detail
      parameters:
        - $ref: "#/components/parameters/OrderId"
      responses:
        "200":
          description: Order detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{orderId}/approval:
    patch:
      operationId: approveBulkOrder
      tags: [orders]
      summary: Accept or reject a bulk order
      parameters:
        - $ref: "#/components/parameters/OrderId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderApprovalInput"
      responses:
        "200":
          description: Updated order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{orderId}/status:
    patch:
      operationId: updateOrderStatus
      tags: [orders]
      summary: Advance delivery tracking
      parameters:
        - $ref: "#/components/parameters/OrderId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderStatusInput"
      responses:
        "200":
          description: Updated order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /reviews:
    post:
      operationId: createReview
      tags: [marketplace]
      summary: Rate a delivered farmer order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewInput"
      responses:
        "201":
          description: Created review
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"
  /farmer/dashboard:
    get:
      operationId: getFarmerDashboard
      tags: [farmers]
      summary: Get farmer dashboard summary
      responses:
        "200":
          description: Dashboard metrics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FarmerDashboard"
  /farmer/recommendations:
    get:
      operationId: getCropRecommendations
      tags: [farmers]
      summary: Get demand-led crop recommendations
      responses:
        "200":
          description: Suggested crops
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CropRecommendation"
  /market-prices/{crop}:
    get:
      operationId: getMarketPrice
      tags: [farmers]
      summary: Get mandi reference price and smart pricing guidance
      parameters:
        - name: crop
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Market reference
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MarketPrice"
components:
  parameters:
    ListingId:
      name: listingId
      in: path
      required: true
      schema:
        type: number
    FarmerId:
      name: farmerId
      in: path
      required: true
      schema:
        type: number
    OrderId:
      name: orderId
      in: path
      required: true
      schema:
        type: number
    CropQuery:
      name: crop
      in: query
      schema:
        type: string
    LocationQuery:
      name: location
      in: query
      schema:
        type: string
    MinPriceQuery:
      name: minPrice
      in: query
      schema:
        type: number
    MaxPriceQuery:
      name: maxPrice
      in: query
      schema:
        type: number
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    Farmer:
      type: object
      required: [id, name, phone, location, rating, reviewCount, bio, initials]
      properties:
        id: { type: number }
        name: { type: string }
        phone: { type: string }
        location: { type: string }
        rating: { type: number }
        reviewCount: { type: number }
        bio: { type: string }
        initials: { type: string }
    Listing:
      type: object
      required: [id, farmerId, farmerName, crop, variety, location, photoUrl, householdPrice, householdQuantity, bulkPrice, bulkQuantity, unit, marketPrice, demandLevel, soldOut]
      properties:
        id: { type: number }
        farmerId: { type: number }
        farmerName: { type: string }
        crop: { type: string }
        variety: { type: string }
        location: { type: string }
        photoUrl: { type: string }
        householdPrice: { type: number }
        householdQuantity: { type: number }
        bulkPrice: { type: number }
        bulkQuantity: { type: number }
        unit: { type: string }
        marketPrice: { type: number }
        demandLevel: { type: string, enum: [high, steady, low] }
        soldOut: { type: boolean }
        fairPrice: { type: boolean }
    ListingInput:
      type: object
      required: [farmerId, crop, variety, location, photoUrl, householdPrice, householdQuantity, bulkPrice, bulkQuantity]
      properties:
        farmerId: { type: number }
        crop: { type: string }
        variety: { type: string }
        location: { type: string }
        photoUrl: { type: string }
        householdPrice: { type: number }
        householdQuantity: { type: number }
        bulkPrice: { type: number }
        bulkQuantity: { type: number }
    ListingUpdate:
      type: object
      properties:
        householdPrice: { type: number }
        householdQuantity: { type: number }
        bulkPrice: { type: number }
        bulkQuantity: { type: number }
        soldOut: { type: boolean }
    Order:
      type: object
      required: [id, trackingId, listingId, crop, farmerName, farmerLocation, buyerName, type, quantity, total, hub, status, stages, createdAt]
      properties:
        id: { type: number }
        trackingId: { type: string }
        listingId: { type: number }
        crop: { type: string }
        farmerName: { type: string }
        farmerLocation: { type: string }
        buyerName: { type: string }
        type: { type: string, enum: [household, bulk] }
        quantity: { type: number }
        total: { type: number }
        hub: { type: string }
        status: { type: string, enum: [order_placed, farmer_confirmed, at_collection_hub, logistics_assigned, out_for_delivery, delivered, rejected] }
        stages:
          type: array
          items:
            $ref: "#/components/schemas/OrderStage"
        createdAt: { type: string }
    OrderStage:
      type: object
      required: [key, label, complete, active]
      properties:
        key: { type: string }
        label: { type: string }
        complete: { type: boolean }
        active: { type: boolean }
    OrderInput:
      type: object
      required: [listingId, buyerName, type, quantity, hub]
      properties:
        listingId: { type: number }
        buyerName: { type: string }
        type: { type: string, enum: [household, bulk] }
        quantity: { type: number }
        hub: { type: string }
    OrderApprovalInput:
      type: object
      required: [decision]
      properties:
        decision: { type: string, enum: [accepted, rejected] }
    OrderStatusInput:
      type: object
      required: [status]
      properties:
        status: { type: string, enum: [farmer_confirmed, at_collection_hub, logistics_assigned, out_for_delivery, delivered] }
    Review:
      type: object
      required: [id, farmerId, buyerName, rating, comment, createdAt]
      properties:
        id: { type: number }
        farmerId: { type: number }
        buyerName: { type: string }
        rating: { type: number, minimum: 1, maximum: 5 }
        comment: { type: string }
        createdAt: { type: string }
    ReviewInput:
      type: object
      required: [orderId, farmerId, buyerName, rating, comment]
      properties:
        orderId: { type: number }
        farmerId: { type: number }
        buyerName: { type: string }
        rating: { type: number, minimum: 1, maximum: 5 }
        comment: { type: string }
    FarmerDashboard:
      type: object
      required: [totalSales, totalProfit, bestSellingCrop, pendingApprovals, chart]
      properties:
        totalSales: { type: number }
        totalProfit: { type: number }
        bestSellingCrop: { type: string }
        pendingApprovals: { type: number }
        chart:
          type: array
          items:
            type: object
            required: [label, value]
            properties:
              label: { type: string }
              value: { type: number }
    CropRecommendation:
      type: object
      required: [crop, demand, reason, score]
      properties:
        crop: { type: string }
        demand: { type: string }
        reason: { type: string }
        score: { type: number }
    MarketPrice:
      type: object
      required: [crop, marketPrice, sustainableLow, sustainableHigh, demandLevel, demandVolume, recommendation, trend]
      properties:
        crop: { type: string }
        marketPrice: { type: number }
        sustainableLow: { type: number }
        sustainableHigh: { type: number }
        demandLevel: { type: string, enum: [high, steady, low] }
        demandVolume: { type: number }
        recommendation: { type: string }
        trend: { type: string, enum: [up, flat, down] }