# lib/api-zod/src/generated/types/cropRecommendation.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * AGRO TRADE marketplace API
 * OpenAPI spec version: 0.1.0
 */

export interface CropRecommendation {
  crop: string;
  demand: string;
  reason: string;
  score: number;
}