# lib/api-zod/src/generated/types/farmerDashboard.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * AGRO TRADE marketplace API
 * OpenAPI spec version: 0.1.0
 */
import type { FarmerDashboardChartItem } from './farmerDashboardChartItem';

export interface FarmerDashboard {
  totalSales: number;
  totalProfit: number;
  bestSellingCrop: string;
  pendingApprovals: number;
  chart: FarmerDashboardChartItem[];
}