# lib/api-zod/src/generated/types/listing.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * AGRO TRADE marketplace API
 * OpenAPI spec version: 0.1.0
 */
import type { ListingDemandLevel } from './listingDemandLevel';

export interface Listing {
  id: number;
  farmerId: number;
  farmerName: string;
  crop: string;
  variety: string;
  location: string;
  photoUrl: string;
  householdPrice: number;
  householdQuantity: number;
  bulkPrice: number;
  bulkQuantity: number;
  unit: string;
  marketPrice: number;
  demandLevel: ListingDemandLevel;
  soldOut: boolean;
  fairPrice?: boolean;
}